# packages/core/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stellar Lense API",
    "description": "Stellar Lense REST API for Stellar wallet risk scoring, alert management, AMM pool risk, cross-chain linking, and compliance exports. See [ReDoc](/redoc) for an alternative documentation view.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/export/scores.csv": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Csv",
        "description": "Stream risk scores as CSV. Max 90-day window. Requires admin key.",
        "operationId": "export_csv_export_scores_csv_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start date YYYY-MM-DD",
              "title": "From"
            },
            "description": "Start date YYYY-MM-DD"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End date YYYY-MM-DD",
              "title": "To"
            },
            "description": "End date YYYY-MM-DD"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "default": 0,
              "title": "Min Score"
            }
          },
          {
            "name": "wallet",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wallet"
            }
          },
          {
            "name": "x-stellarlense-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-StellarLense-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/export/scores.parquet": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Parquet",
        "description": "Stream risk scores as Parquet (snappy compressed). Max 90-day window. Requires admin key.",
        "operationId": "export_parquet_export_scores_parquet_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start date YYYY-MM-DD",
              "title": "From"
            },
            "description": "Start date YYYY-MM-DD"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End date YYYY-MM-DD",
              "title": "To"
            },
            "description": "End date YYYY-MM-DD"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "default": 0,
              "title": "Min Score"
            }
          },
          {
            "name": "wallet",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wallet"
            }
          },
          {
            "name": "x-stellarlense-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-StellarLense-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/audit/wallet/{wallet}": {
      "get": {
        "tags": [
          "Audit Log"
        ],
        "summary": "Get scoring event history for a wallet",
        "description": "Returns the full chronological scoring event history for the given wallet address. Requires admin API key.",
        "operationId": "get_wallet_audit_log_audit_wallet__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "x-stellarlense-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-StellarLense-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoringEventResponse"
                  },
                  "title": "Response Get Wallet Audit Log Audit Wallet  Wallet  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit/wallet/{wallet}/verify": {
      "get": {
        "tags": [
          "Audit Log"
        ],
        "summary": "Verify the audit chain integrity for a wallet",
        "description": "Walks the scoring event chain for the wallet, recomputing each chain hash. Returns VALID, TAMPERED (with first failing event ID), or no_events.",
        "operationId": "verify_wallet_chain_audit_wallet__wallet__verify_get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          },
          {
            "name": "x-stellarlense-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-StellarLense-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit/summary": {
      "get": {
        "tags": [
          "Audit Log"
        ],
        "summary": "Get audit log summary statistics",
        "description": "Returns aggregate statistics for the audit log: events in the last 24 hours, unique wallets scored, and integrity violations detected.",
        "operationId": "get_audit_summary_audit_summary_get",
        "parameters": [
          {
            "name": "x-stellarlense-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-StellarLense-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditSummaryResponse": {
        "properties": {
          "events_last_24h": {
            "type": "integer",
            "title": "Events Last 24H"
          },
          "unique_wallets_scored": {
            "type": "integer",
            "title": "Unique Wallets Scored"
          },
          "integrity_violations": {
            "type": "integer",
            "title": "Integrity Violations"
          }
        },
        "type": "object",
        "required": [
          "events_last_24h",
          "unique_wallets_scored",
          "integrity_violations"
        ],
        "title": "AuditSummaryResponse"
      },
...
      "ChainVerificationResponse": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_events": {
            "type": "integer",
            "title": "Total Events"
          },
          "first_tampered_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Tampered Event Id"
          },
          "verified_at": {
            "type": "string",
            "title": "Verified At"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "status",
          "total_events",
          "first_tampered_event_id",
          "verified_at"
        ],
        "title": "ChainVerificationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScoringEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "namespace_id": {
            "type": "string",
            "title": "Namespace Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "previous_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Score"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "triggered_by": {
            "type": "string",
            "title": "Triggered By"
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "chain_hash": {
            "type": "string",
            "title": "Chain Hash"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "feature_snapshot_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Feature Snapshot Keys"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "wallet",
          "namespace_id",
          "score",
          "previous_score",
          "model_version",
          "triggered_by",
          "actor_id",
          "chain_hash",
          "occurred_at",
          "feature_snapshot_keys"
        ],
        "title": "ScoringEventResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}